Migrate compute resources out of terraform resource management
