+++ deployment-manager/matlab-flexnet/matlab-flexnet.yaml
@@ -1,0 +1,46 @@
+# Copyright 2017 SchedMD LLC.
+# Modified for use with the Slurm Resource Manager.
+#
+# Copyright 2015 Google Inc. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+# [START cluster_yaml]
+imports:
+- path: matlab-flexnet.jinja
+
+resources:
+- name: matlab-flexnet-license-server
+  type: matlab-flexnet.jinja
+  properties:
+
+    # ///////////////////////////////////////////////// #
+    #               Host Project Settings               #
+    vpc_net                 : bsc-host-network
+    shared_vpc_host_proj    : pennbrain-host-3097383fff
+    # ///////////////////////////////////////////////// #
+    
+    # ///////////////////////////////////////////////// #
+    #              Service Project Settings             #
+    name                    : matlab-flexnet
+    vpc_subnet              : pennbrain-license-subnet
+    zone                    : us-east1-b
+    region                  : us-east1
+    network_tag             : pennbrain-license
+
+    # Fileserver Settings
+    machine_type      : n1-standard-4
+    boot_disk_type    : pd-standard
+    boot_disk_size_gb : 50
+             
+#  [END cluster_yaml]
