+++ modules/host_project/variables.tf
@@ -23,7 +23,7 @@ variable "network_name"{
   type = string
 }
 
-variable "project_id_base"{
+variable "project_id"{
   type = string
 }
