+++ resources/.terraform/modules/modules.json
@@ -1,0 +1,1 @@
+{"Modules":[{"Key":"license_project","Source":"../modules/license_project","Dir":"../modules/license_project"},{"Key":"user_projects.zfs-file-server","Source":"./storage/zfs_fileserver","Dir":"../modules/user_projects/storage/zfs_fileserver"},{"Key":"service_project","Source":"../modules/service_project","Dir":"../modules/service_project"},{"Key":"service_projects","Source":"../modules/service_projects","Dir":"../modules/service_projects"},{"Key":"user_projects","Source":"../modules/user_projects","Dir":""},{"Key":"user_projects.network_resources","Source":"./network","Dir":"../modules/user_projects/network"},{"Key":"","Source":"","Dir":"."},{"Key":"host_project","Source":"../modules/host_project","Dir":"../modules/host_project"}]}
