+++ resources/.terraform/plugins/linux_amd64/lock.json
@@ -1,0 +1,3 @@
+{
+  "google": "fc4fa849aa954dc75b47db9526f9b989fbd0b6232b5cc1509c6e850946dd7e3d"
+}
