+++ resources/template.tfvars
@@ -1,37 +1,30 @@
 credentials_path      = "TO DO : Enter the full path to the credentials file for your terraform service account"
 parent_resource_node  = "TO DO : Enter the resource hierarchy node that host the projects and folders created from these scripts."
 billing_account       = "TO DO"
+services_folder       = "pennbrain-services"
 
 host_project={
   name = "host-project"
   network_name = "bsc-host-network"
-  parent_folder = "services"
-  project_id_base = "bsc-host"
+  project_id   = "pennbrain-host-3097383fff"
 }
 
-license_project = {
-  name                 = "license"
-  network_resources    = {
-    subnet_cidr      = "10.10.0.0/16"
-    subnet_region    = "us-east1"
-    source_ranges    = ["10.20.0.0/16"]
-    allow            = [
-    { protocol = "tcp"
-      ports    = ["27000"]
-    }]
-  }
-  compute_resources = [
-    { server_name  = "matlab-flexnet"
-      machine_type = "n1-standard-2"
-      disk_type    = "pd-standard"
-      disk_size_gb = 25
-      zone         = "us-east1-b"
+service_projects = [
+  { name                 = "pennbrain-license"
+    project_id           = "pennbrain-license-82eff5"
+    network_resources    = {
+      subnet_cidr      = "10.10.0.0/16"
+      subnet_region    = "us-east1"
+      source_ranges    = ["10.20.0.0/16"]
+      allow            = [
+      { protocol = "tcp"
+        ports    = ["27000","27001","1049"]
+      }]
     }
-  ]
-}
-
-user_projects = [
+    members = ["group:admin@gcp.pennbrain.upenn.edu"]
+  },
   { name            = "detre-group"
+    project_id      = "detre-group-dd34a9"
     network_resources = {
       subnet_cidr = "10.20.0.0/16"
       subnet_region = "us-east1"
@@ -41,34 +34,6 @@ user_projects = [
         ports    = ["22"]
       }]
     }
-    storage_resources = {
-      server_name  = "zfs-fileserver"
-      machine_type = "n1-standard-4"
-      disk_name    = "zfs-fs-storage"
-      disk_type    = "pd-standard"
-      disk_size_gb = 20000
-      zone         = "us-east1-b"
-    }
     members = ["group:detre-group@gcp.pennbrain.upenn.edu"]
-  },
-  { name            = "another-group"
-    network_resources = {
-      subnet_cidr = "10.30.0.0/16"
-      subnet_region = "us-east1"
-      source_ranges = ["0.0.0.0/0"]
-      allow = [
-      { protocol = "tcp"
-        ports    = ["22"]
-      }]
-    }
-    storage_resources = {
-      server_name  = "zfs-fileserver"
-      machine_type = "n1-standard-4"
-      disk_name    = "zfs-fs-storage"
-      disk_type    = "pd-standard"
-      disk_size_gb = 10000
-      zone         = "us-east1-c"
-    }
-    members = ["group:another-group@gcp.pennbrain.upenn.edu","user:joe@gcp.pennbrain.upenn.edu","org:fluidnumerics.com"]
   }
-}]
+]
