+++ resources/variables.tf
@@ -18,35 +18,14 @@ variable "host_project"{
   type = object({
     name                 = string
     network_name         = string
-    project_id_base      = string
+    project_id           = string
   })
 }
 
-variable "license_project"{
-  type = object({
-    name                 = string
-    network_resources    = object({
-      subnet_cidr      = string
-      subnet_region    = string
-      source_ranges    = list( string )
-      allow            = list(object({
-        protocol = string
-        ports    = list(string)
-      }))
-    })
-    compute_resources = list(object({
-      server_name  = string
-      machine_type = string
-      disk_type    = string
-      disk_size_gb = number
-      zone         = string
-    }))
-  })
-}
-
-variable "user_projects"{
+variable "service_projects"{
   type = list(object({
     name                 = string
+    project_id           = string
     network_resources    = object({
       subnet_cidr      = string
       subnet_region    = string
@@ -55,14 +34,6 @@ variable "user_projects"{
         protocol = string
         ports    = list(string)
       }))
-    })
-    storage_resources = object({
-      server_name  = string
-      machine_type = string
-      disk_name    = string
-      disk_type    = string
-      disk_size_gb = number
-      zone         = string
     })
     members = list(string)
   }))
